Add STM32 HAL and stm32pintester board config with bootloader offset
Adds STM32/maple HAL layer (XInput, GPIO, serial, keyboard mode) and
stm32pintester board configuration. App is linked at 0x08001000 with
VECT_TAB_ADDR relocated to support a 4KB custom bootloader at 0x08000000.

+++ src/util/analog_filters.cpp
@@ -14,7 +14,7 @@ uint8_t apply_deadzone(uint8_t value, uint8_t deadzone, bool scale) {
         // value is given on the rim.
         if (scale) {
             int8_t sign = SIGNUM(post_deadzone);
-            int8_t post_scaling = min(127, abs(post_deadzone) * 128.0 / (128 - deadzone)) * sign;
+            int8_t post_scaling = min(127, (int)(abs(post_deadzone) * 128.0 / (128 - deadzone))) * sign;
             return post_scaling + 128;
         }
         return post_deadzone + 128;
